refactor: divide Request and Response in Dto Package
