+++ src/main/java/com/kuit/kuit4serverauth/DTO/response/LoginSuccessResponse.java
@@ -1,0 +1,11 @@
+package com.kuit.kuit4serverauth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class LoginSuccessResponse {
+    private String accessToken;
+    //private String refreshToken;
+}
